refactor(host-application): extract app_service tests from mod.rs
Move the oversized inline app_service #[cfg(test)] module into dedicated files under app_service/tests.

Keep mod.rs focused on runtime module wiring and split tests into shared fixtures, unit coverage, and integration coverage.
