+++ apps/desktop/src-tauri/crates/host-application/src/app_service/tests/mod.rs
@@ -1,0 +1,895 @@
+use super::build_orchestrator::{BuildResponseAction, CleanupMode};
+use super::{
+    allows_transition, build_opencode_config_content, build_opencode_startup_event_payload,
+    can_set_plan, can_set_spec_from_status, default_mcp_workspace_root, derive_available_actions,
+    is_orphaned_opencode_server_process, normalize_required_markdown,
+    normalize_subtask_plan_inputs, parse_mcp_command_json, read_opencode_process_registry,
+    read_opencode_version, resolve_mcp_command, resolve_opencode_binary_path,
+    terminate_child_process, terminate_process_by_pid, validate_parent_relationships_for_create,
+    validate_parent_relationships_for_update, validate_plan_subtask_rules, validate_transition,
+    wait_for_local_server, wait_for_local_server_with_process,
+    with_locked_opencode_process_registry, AgentRuntimeProcess, AppService,
+    OpencodeProcessRegistryInstance, OpencodeStartupMetricsSnapshot,
+    OpencodeStartupReadinessPolicy, OpencodeStartupWaitReport, RunProcess,
+    TrackedOpencodeProcessGuard, OPENCODE_PROCESS_REGISTRY_RELATIVE_PATH,
+};
+use anyhow::{anyhow, Context, Result};
+use host_domain::{
+    AgentRuntimeSummary, AgentSessionDocument, AgentWorkflows, CreateTaskInput, GitBranch,
+    GitCurrentBranch, GitPort, GitPushSummary, PlanSubtaskInput, QaReportDocument, QaVerdict,
+    RunEvent, RunState, RunSummary, SpecDocument, TaskAction, TaskCard, TaskDocumentSummary,
+    TaskMetadata, TaskStatus, TaskStore, UpdateTaskPatch,
+};
+use host_infra_system::{
+    AppConfigStore, GlobalConfig, HookSet, OpencodeStartupReadinessConfig, RepoConfig,
+};
+use serde_json::Value;
+use std::ffi::OsString;
+use std::fs;
+use std::io::Write;
+use std::net::{TcpListener, TcpStream};
+use std::path::{Path, PathBuf};
+use std::process::{Command, Stdio};
+use std::sync::atomic::{AtomicU64, Ordering};
+use std::sync::{Arc, LazyLock, Mutex};
+use std::time::{Duration, Instant, SystemTime, UNIX_EPOCH};
+
+fn make_task(id: &str, issue_type: &str, status: TaskStatus) -> TaskCard {
+    TaskCard {
+        id: id.to_string(),
+        title: format!("Task {id}"),
+        description: String::new(),
+        acceptance_criteria: String::new(),
+        notes: String::new(),
+        status,
+        priority: 2,
+        issue_type: issue_type.to_string(),
+        ai_review_enabled: true,
+        available_actions: Vec::new(),
+        labels: Vec::new(),
+        assignee: None,
+        parent_id: None,
+        subtask_ids: Vec::new(),
+        document_summary: TaskDocumentSummary::default(),
+        agent_workflows: AgentWorkflows::default(),
+        updated_at: "2026-01-01T00:00:00Z".to_string(),
+        created_at: "2026-01-01T00:00:00Z".to_string(),
+    }
+}
+
+#[derive(Debug, Default)]
+struct TaskStoreState {
+    ensure_calls: Vec<String>,
+    ensure_error: Option<String>,
+    tasks: Vec<TaskCard>,
+    list_error: Option<String>,
+    delete_calls: Vec<(String, bool)>,
+    created_inputs: Vec<CreateTaskInput>,
+    updated_patches: Vec<(String, UpdateTaskPatch)>,
+    spec_get_calls: Vec<String>,
+    spec_set_calls: Vec<(String, String)>,
+    plan_get_calls: Vec<String>,
+    plan_set_calls: Vec<(String, String)>,
+    metadata_get_calls: Vec<String>,
+    qa_append_calls: Vec<(String, String, QaVerdict)>,
+    latest_qa_report: Option<QaReportDocument>,
+    agent_sessions: Vec<AgentSessionDocument>,
+    upserted_sessions: Vec<(String, AgentSessionDocument)>,
+}
+
+#[derive(Clone)]
+struct FakeTaskStore {
+    state: Arc<Mutex<TaskStoreState>>,
+}
+
+impl TaskStore for FakeTaskStore {
+    fn ensure_repo_initialized(&self, repo_path: &Path) -> Result<()> {
+        let mut state = self.state.lock().expect("task store lock poisoned");
+        if let Some(message) = state.ensure_error.as_ref() {
+            return Err(anyhow!(message.clone()));
+        }
+        state
+            .ensure_calls
+            .push(repo_path.to_string_lossy().to_string());
+        Ok(())
+    }
+
+    fn list_tasks(&self, _repo_path: &Path) -> Result<Vec<TaskCard>> {
+        let state = self.state.lock().expect("task store lock poisoned");
+        if let Some(message) = state.list_error.as_ref() {
+            return Err(anyhow!(message.clone()));
+        }
+        Ok(state.tasks.clone())
+    }
+
+    fn create_task(&self, _repo_path: &Path, input: CreateTaskInput) -> Result<TaskCard> {
+        let mut state = self.state.lock().expect("task store lock poisoned");
+        state.created_inputs.push(input.clone());
+        let task = TaskCard {
+            id: format!("generated-{}", state.tasks.len() + 1),
+            title: input.title,
+            description: input.description.unwrap_or_default(),
+            acceptance_criteria: input.acceptance_criteria.unwrap_or_default(),
+            notes: String::new(),
+            status: TaskStatus::Open,
+            priority: input.priority,
+            issue_type: input.issue_type,
+            ai_review_enabled: input.ai_review_enabled.unwrap_or(true),
+            available_actions: Vec::new(),
+            labels: input.labels.unwrap_or_default(),
+            assignee: None,
+            parent_id: input.parent_id,
+            subtask_ids: Vec::new(),
+            document_summary: TaskDocumentSummary::default(),
+            agent_workflows: AgentWorkflows::default(),
+            updated_at: "2026-01-01T00:00:00Z".to_string(),
+            created_at: "2026-01-01T00:00:00Z".to_string(),
+        };
+        state.tasks.push(task.clone());
+        Ok(task)
+    }
+
+    fn update_task(
+        &self,
+        _repo_path: &Path,
+        task_id: &str,
+        patch: UpdateTaskPatch,
+    ) -> Result<TaskCard> {
+        let mut state = self.state.lock().expect("task store lock poisoned");
+        state
+            .updated_patches
+            .push((task_id.to_string(), patch.clone()));
+        let index = state
+            .tasks
+            .iter()
+            .position(|task| task.id == task_id)
+            .ok_or_else(|| anyhow!("task not found: {task_id}"))?;
+
+        let mut updated = state.tasks[index].clone();
+        if let Some(title) = patch.title {
+            updated.title = title;
+        }
+        if let Some(status) = patch.status {
+            updated.status = status;
+        }
+        if let Some(issue_type) = patch.issue_type {
+            updated.issue_type = issue_type;
+        }
+        if let Some(ai_review_enabled) = patch.ai_review_enabled {
+            updated.ai_review_enabled = ai_review_enabled;
+        }
+        if let Some(parent_id) = patch.parent_id {
+            updated.parent_id = Some(parent_id);
+        }
+        if let Some(labels) = patch.labels {
+            updated.labels = labels;
+        }
+
+        state.tasks[index] = updated.clone();
+        Ok(updated)
+    }
+
+    fn delete_task(&self, _repo_path: &Path, task_id: &str, delete_subtasks: bool) -> Result<bool> {
+        let mut state = self.state.lock().expect("task store lock poisoned");
+        state
+            .delete_calls
+            .push((task_id.to_string(), delete_subtasks));
+        Ok(true)
+    }
+
+    fn get_spec(&self, _repo_path: &Path, _task_id: &str) -> Result<SpecDocument> {
+        let mut state = self.state.lock().expect("task store lock poisoned");
+        state.spec_get_calls.push(_task_id.to_string());
+        Ok(SpecDocument {
+            markdown: String::new(),
+            updated_at: None,
+        })
+    }
+
+    fn set_spec(&self, _repo_path: &Path, _task_id: &str, markdown: &str) -> Result<SpecDocument> {
+        let mut state = self.state.lock().expect("task store lock poisoned");
+        state
+            .spec_set_calls
+            .push((_task_id.to_string(), markdown.to_string()));
+        Ok(SpecDocument {
+            markdown: markdown.to_string(),
+            updated_at: Some("2026-01-01T00:00:00Z".to_string()),
+        })
+    }
+
+    fn get_plan(&self, _repo_path: &Path, _task_id: &str) -> Result<SpecDocument> {
+        let mut state = self.state.lock().expect("task store lock poisoned");
+        state.plan_get_calls.push(_task_id.to_string());
+        Ok(SpecDocument {
+            markdown: String::new(),
+            updated_at: None,
+        })
+    }
+
+    fn set_plan(&self, _repo_path: &Path, _task_id: &str, markdown: &str) -> Result<SpecDocument> {
+        let mut state = self.state.lock().expect("task store lock poisoned");
+        state
+            .plan_set_calls
+            .push((_task_id.to_string(), markdown.to_string()));
+        Ok(SpecDocument {
+            markdown: markdown.to_string(),
+            updated_at: Some("2026-01-01T00:00:00Z".to_string()),
+        })
+    }
+
+    fn get_latest_qa_report(
+        &self,
+        _repo_path: &Path,
+        _task_id: &str,
+    ) -> Result<Option<QaReportDocument>> {
+        let state = self.state.lock().expect("task store lock poisoned");
+        Ok(state.latest_qa_report.clone())
+    }
+
+    fn append_qa_report(
+        &self,
+        _repo_path: &Path,
+        _task_id: &str,
+        markdown: &str,
+        verdict: QaVerdict,
+    ) -> Result<QaReportDocument> {
+        let mut state = self.state.lock().expect("task store lock poisoned");
+        state
+            .qa_append_calls
+            .push((_task_id.to_string(), markdown.to_string(), verdict.clone()));
+        Ok(QaReportDocument {
+            markdown: markdown.to_string(),
+            verdict,
+            updated_at: "2026-01-01T00:00:00Z".to_string(),
+            revision: 1,
+        })
+    }
+
+    fn list_agent_sessions(
+        &self,
+        _repo_path: &Path,
+        _task_id: &str,
+    ) -> Result<Vec<AgentSessionDocument>> {
+        let state = self.state.lock().expect("task store lock poisoned");
+        Ok(state.agent_sessions.clone())
+    }
+
+    fn upsert_agent_session(
+        &self,
+        _repo_path: &Path,
+        _task_id: &str,
+        session: AgentSessionDocument,
+    ) -> Result<()> {
+        let mut state = self.state.lock().expect("task store lock poisoned");
+        state
+            .upserted_sessions
+            .push((_task_id.to_string(), session.clone()));
+        if let Some(index) = state
+            .agent_sessions
+            .iter()
+            .position(|entry| entry.session_id == session.session_id)
+        {
+            state.agent_sessions[index] = session;
+        } else {
+            state.agent_sessions.push(session);
+        }
+        Ok(())
+    }
+
+    fn get_task_metadata(&self, _repo_path: &Path, _task_id: &str) -> Result<TaskMetadata> {
+        let mut state = self.state.lock().expect("task store lock poisoned");
+        state.metadata_get_calls.push(_task_id.to_string());
+        let qa_report = state.latest_qa_report.clone();
+        let agent_sessions = state.agent_sessions.clone();
+        Ok(TaskMetadata {
+            spec: SpecDocument {
+                markdown: String::new(),
+                updated_at: None,
+            },
+            plan: SpecDocument {
+                markdown: String::new(),
+                updated_at: None,
+            },
+            qa_report,
+            agent_sessions,
+        })
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+enum GitCall {
+    GetBranches {
+        repo_path: String,
+    },
+    GetCurrentBranch {
+        repo_path: String,
+    },
+    SwitchBranch {
+        repo_path: String,
+        branch: String,
+        create: bool,
+    },
+    CreateWorktree {
+        repo_path: String,
+        worktree_path: String,
+        branch: String,
+        create_branch: bool,
+    },
+    RemoveWorktree {
+        repo_path: String,
+        worktree_path: String,
+        force: bool,
+    },
+    PushBranch {
+        repo_path: String,
+        remote: String,
+        branch: String,
+        set_upstream: bool,
+        force_with_lease: bool,
+    },
+}
+
+#[derive(Debug)]
+struct GitState {
+    calls: Vec<GitCall>,
+    branches: Vec<GitBranch>,
+    current_branch: GitCurrentBranch,
+}
+
+#[derive(Clone)]
+struct FakeGitPort {
+    state: Arc<Mutex<GitState>>,
+}
+
+impl GitPort for FakeGitPort {
+    fn get_branches(&self, repo_path: &Path) -> Result<Vec<GitBranch>> {
+        let mut state = self.state.lock().expect("git state lock poisoned");
+        state.calls.push(GitCall::GetBranches {
+            repo_path: repo_path.to_string_lossy().to_string(),
+        });
+        Ok(state.branches.clone())
+    }
+
+    fn get_current_branch(&self, repo_path: &Path) -> Result<GitCurrentBranch> {
+        let mut state = self.state.lock().expect("git state lock poisoned");
+        state.calls.push(GitCall::GetCurrentBranch {
+            repo_path: repo_path.to_string_lossy().to_string(),
+        });
+        Ok(state.current_branch.clone())
+    }
+
+    fn switch_branch(
+        &self,
+        repo_path: &Path,
+        branch: &str,
+        create: bool,
+    ) -> Result<GitCurrentBranch> {
+        let mut state = self.state.lock().expect("git state lock poisoned");
+        state.calls.push(GitCall::SwitchBranch {
+            repo_path: repo_path.to_string_lossy().to_string(),
+            branch: branch.to_string(),
+            create,
+        });
+        state.current_branch = GitCurrentBranch {
+            name: Some(branch.to_string()),
+            detached: false,
+        };
+        Ok(state.current_branch.clone())
+    }
+
+    fn create_worktree(
+        &self,
+        repo_path: &Path,
+        worktree_path: &Path,
+        branch: &str,
+        create_branch: bool,
+    ) -> Result<()> {
+        let mut state = self.state.lock().expect("git state lock poisoned");
+        state.calls.push(GitCall::CreateWorktree {
+            repo_path: repo_path.to_string_lossy().to_string(),
+            worktree_path: worktree_path.to_string_lossy().to_string(),
+            branch: branch.to_string(),
+            create_branch,
+        });
+        Ok(())
+    }
+
+    fn remove_worktree(&self, repo_path: &Path, worktree_path: &Path, force: bool) -> Result<()> {
+        let mut state = self.state.lock().expect("git state lock poisoned");
+        state.calls.push(GitCall::RemoveWorktree {
+            repo_path: repo_path.to_string_lossy().to_string(),
+            worktree_path: worktree_path.to_string_lossy().to_string(),
+            force,
+        });
+        Ok(())
+    }
+
+    fn push_branch(
+        &self,
+        repo_path: &Path,
+        remote: &str,
+        branch: &str,
+        set_upstream: bool,
+        force_with_lease: bool,
+    ) -> Result<GitPushSummary> {
+        let mut state = self.state.lock().expect("git state lock poisoned");
+        state.calls.push(GitCall::PushBranch {
+            repo_path: repo_path.to_string_lossy().to_string(),
+            remote: remote.to_string(),
+            branch: branch.to_string(),
+            set_upstream,
+            force_with_lease,
+        });
+        Ok(GitPushSummary {
+            remote: remote.to_string(),
+            branch: branch.to_string(),
+            output: "ok".to_string(),
+        })
+    }
+}
+
+fn build_service_with_state(
+    tasks: Vec<TaskCard>,
+    branches: Vec<GitBranch>,
+    current_branch: GitCurrentBranch,
+) -> (AppService, Arc<Mutex<TaskStoreState>>, Arc<Mutex<GitState>>) {
+    let task_state = Arc::new(Mutex::new(TaskStoreState {
+        ensure_calls: Vec::new(),
+        ensure_error: None,
+        tasks,
+        list_error: None,
+        delete_calls: Vec::new(),
+        created_inputs: Vec::new(),
+        updated_patches: Vec::new(),
+        spec_get_calls: Vec::new(),
+        spec_set_calls: Vec::new(),
+        plan_get_calls: Vec::new(),
+        plan_set_calls: Vec::new(),
+        metadata_get_calls: Vec::new(),
+        qa_append_calls: Vec::new(),
+        latest_qa_report: None,
+        agent_sessions: Vec::new(),
+        upserted_sessions: Vec::new(),
+    }));
+    let git_state = Arc::new(Mutex::new(GitState {
+        calls: Vec::new(),
+        branches,
+        current_branch,
+    }));
+    let task_store: Arc<dyn TaskStore> = Arc::new(FakeTaskStore {
+        state: task_state.clone(),
+    });
+    let git_port: Arc<dyn GitPort> = Arc::new(FakeGitPort {
+        state: git_state.clone(),
+    });
+    let config_store = AppConfigStore::from_path(unique_temp_path("host-app-config"));
+    let service = AppService::with_git_port(task_store, config_store, git_port);
+    (service, task_state, git_state)
+}
+
+fn make_session(task_id: &str, session_id: &str) -> AgentSessionDocument {
+    AgentSessionDocument {
+        session_id: session_id.to_string(),
+        external_session_id: format!("external-{session_id}"),
+        task_id: task_id.to_string(),
+        role: "build".to_string(),
+        scenario: "build_default".to_string(),
+        status: "running".to_string(),
+        started_at: "2026-02-20T12:00:00Z".to_string(),
+        updated_at: "2026-02-20T12:00:10Z".to_string(),
+        ended_at: None,
+        runtime_id: Some("runtime-1".to_string()),
+        run_id: Some("run-1".to_string()),
+        base_url: "http://127.0.0.1:4173".to_string(),
+        working_directory: "/tmp/repo".to_string(),
+        selected_model: None,
+    }
+}
+
+static ENV_LOCK: LazyLock<Mutex<()>> = LazyLock::new(|| Mutex::new(()));
+
+fn lock_env<'a>() -> std::sync::MutexGuard<'a, ()> {
+    ENV_LOCK.lock().unwrap_or_else(|poison| poison.into_inner())
+}
+
+fn unique_temp_path(name: &str) -> PathBuf {
+    let nonce = SystemTime::now()
+        .duration_since(UNIX_EPOCH)
+        .expect("system time")
+        .as_nanos();
+    std::env::temp_dir().join(format!("openducktor-host-app-{name}-{nonce}"))
+}
+
+fn write_executable_script(path: &Path, script: &str) -> Result<()> {
+    if let Some(parent) = path.parent() {
+        fs::create_dir_all(parent)?;
+    }
+    let mut file = fs::File::create(path)?;
+    file.write_all(script.as_bytes())?;
+    let status = Command::new("chmod")
+        .arg("+x")
+        .arg(path)
+        .status()
+        .map_err(|error| anyhow!("failed running chmod: {error}"))?;
+    if !status.success() {
+        return Err(anyhow!("chmod +x failed for {}", path.display()));
+    }
+    Ok(())
+}
+
+fn init_git_repo(path: &Path) -> Result<()> {
+    fs::create_dir_all(path)?;
+    Command::new("git")
+        .arg("init")
+        .arg(path)
+        .stdout(Stdio::null())
+        .stderr(Stdio::null())
+        .status()?;
+    Command::new("git")
+        .arg("-C")
+        .arg(path)
+        .arg("config")
+        .arg("user.email")
+        .arg("odt-test@example.com")
+        .status()?;
+    Command::new("git")
+        .arg("-C")
+        .arg(path)
+        .arg("config")
+        .arg("user.name")
+        .arg("OpenDucktor Test")
+        .status()?;
+    fs::write(path.join("README.md"), "# test\n")?;
+    Command::new("git")
+        .arg("-C")
+        .arg(path)
+        .arg("add")
+        .arg(".")
+        .status()?;
+    Command::new("git")
+        .arg("-C")
+        .arg(path)
+        .arg("commit")
+        .arg("-m")
+        .arg("initial")
+        .stdout(Stdio::null())
+        .stderr(Stdio::null())
+        .status()?;
+    Ok(())
+}
+
+fn create_fake_opencode(path: &Path) -> Result<()> {
+    let script = r#"#!/bin/sh
+if [ "$1" = "--version" ]; then
+  echo "opencode-fake 0.0.1"
+  exit 0
+fi
+
+if [ "$1" = "serve" ]; then
+  HOST="127.0.0.1"
+  PORT="0"
+  while [ $# -gt 0 ]; do
+    case "$1" in
+      --hostname)
+        HOST="$2"
+        shift 2
+        ;;
+      --port)
+        PORT="$2"
+        shift 2
+        ;;
+      *)
+        shift
+        ;;
+    esac
+  done
+  echo "permission requested: git push"
+  echo "tool execution heartbeat" >&2
+  exec python3 - "$HOST" "$PORT" <<'PY'
+import os
+import signal
+import socket
+import sys
+import time
+
+host = sys.argv[1]
+port = int(sys.argv[2])
+delay_ms = int(os.environ.get("OPENDUCKTOR_TEST_STARTUP_DELAY_MS", "0") or "0")
+pid_file = os.environ.get("OPENDUCKTOR_TEST_PID_FILE", "")
+termination_file = os.environ.get("OPENDUCKTOR_TEST_TERM_FILE", "")
+
+if pid_file:
+    try:
+        with open(pid_file, "w", encoding="utf-8") as file:
+            file.write(str(os.getpid()))
+    except Exception:
+        pass
+
+if delay_ms > 0:
+    time.sleep(delay_ms / 1000.0)
+
+server = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+server.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
+server.bind((host, port))
+server.listen(16)
+
+def _stop(*_):
+    try:
+        if termination_file:
+            try:
+                with open(termination_file, "w", encoding="utf-8") as file:
+                    file.write("terminated")
+            except Exception:
+                pass
+        server.close()
+    finally:
+        raise SystemExit(0)
+
+signal.signal(signal.SIGTERM, _stop)
+signal.signal(signal.SIGINT, _stop)
+
+while True:
+    conn, _ = server.accept()
+    try:
+        conn.recv(1024)
+    except Exception:
+        pass
+    finally:
+        conn.close()
+PY
+fi
+
+echo "unsupported opencode invocation" >&2
+exit 1
+"#;
+    write_executable_script(path, script)
+}
+
+fn create_failing_opencode(path: &Path) -> Result<()> {
+    let script = r#"#!/bin/sh
+if [ "$1" = "--version" ]; then
+  echo "opencode-fake 0.0.1"
+  exit 0
+fi
+
+if [ "$1" = "serve" ]; then
+  echo "simulated startup failure" >&2
+  exit 42
+fi
+
+echo "unsupported opencode invocation" >&2
+exit 1
+"#;
+    write_executable_script(path, script)
+}
+
+fn create_orphanable_opencode(path: &Path) -> Result<()> {
+    let script = r#"#!/bin/sh
+if [ "$1" = "--version" ]; then
+  echo "opencode-fake 0.0.1"
+  exit 0
+fi
+
+if [ "$1" = "serve" ]; then
+  while true; do
+    sleep 1
+  done
+fi
+
+echo "unsupported opencode invocation" >&2
+exit 1
+"#;
+    write_executable_script(path, script)
+}
+
+fn create_failing_opencode_with_worktree_cleanup(path: &Path) -> Result<()> {
+    let script = r#"#!/bin/sh
+if [ "$1" = "--version" ]; then
+  echo "opencode-fake 0.0.1"
+  exit 0
+fi
+
+if [ "$1" = "serve" ]; then
+  REPO_PATH=$(python3 - <<'PY'
+import json
+import os
+
+raw = os.environ.get("OPENCODE_CONFIG_CONTENT", "{}")
+try:
+    config = json.loads(raw)
+except Exception:
+    print("")
+    raise SystemExit(0)
+
+environment = (
+    config.get("mcp", {})
+    .get("openducktor", {})
+    .get("environment", {})
+)
+print(environment.get("ODT_REPO_PATH", ""))
+PY
+)
+  if [ -n "$REPO_PATH" ]; then
+    rm -rf "$REPO_PATH"
+  fi
+  echo "simulated startup failure after deleting repo path" >&2
+  exit 42
+fi
+
+echo "unsupported opencode invocation" >&2
+exit 1
+"#;
+    write_executable_script(path, script)
+}
+
+fn create_fake_bd(path: &Path) -> Result<()> {
+    let script = r#"#!/bin/sh
+echo "bd-fake"
+"#;
+    write_executable_script(path, script)
+}
+
+struct EnvVarGuard {
+    key: String,
+    previous: Option<OsString>,
+}
+
+impl Drop for EnvVarGuard {
+    fn drop(&mut self) {
+        if let Some(previous) = self.previous.clone() {
+            std::env::set_var(self.key.as_str(), previous);
+        } else {
+            std::env::remove_var(self.key.as_str());
+        }
+    }
+}
+
+fn set_env_var(key: &str, value: &str) -> EnvVarGuard {
+    let previous = std::env::var_os(key);
+    std::env::set_var(key, value);
+    EnvVarGuard {
+        key: key.to_string(),
+        previous,
+    }
+}
+
+fn remove_env_var(key: &str) -> EnvVarGuard {
+    let previous = std::env::var_os(key);
+    std::env::remove_var(key);
+    EnvVarGuard {
+        key: key.to_string(),
+        previous,
+    }
+}
+
+fn prepend_path(path_prefix: &Path) -> EnvVarGuard {
+    let previous = std::env::var_os("PATH");
+    let mut parts = vec![path_prefix.to_string_lossy().to_string()];
+    if let Some(current) = previous.as_ref() {
+        parts.push(current.to_string_lossy().to_string());
+    }
+    let value = parts.join(":");
+    std::env::set_var("PATH", value);
+    EnvVarGuard {
+        key: "PATH".to_string(),
+        previous,
+    }
+}
+
+fn wait_for_path_exists(path: &Path, timeout: Duration) -> bool {
+    let deadline = Instant::now() + timeout;
+    while Instant::now() < deadline {
+        if path.exists() {
+            return true;
+        }
+        std::thread::sleep(Duration::from_millis(50));
+    }
+    path.exists()
+}
+
+fn process_is_alive(pid: i32) -> bool {
+    Command::new("/bin/sh")
+        .arg("-lc")
+        .arg(format!("kill -0 {pid}"))
+        .stdout(Stdio::null())
+        .stderr(Stdio::null())
+        .status()
+        .map(|status| status.success())
+        .unwrap_or(false)
+}
+
+fn wait_for_process_exit(pid: i32, timeout: Duration) -> bool {
+    let deadline = Instant::now() + timeout;
+    while Instant::now() < deadline {
+        if !process_is_alive(pid) {
+            return true;
+        }
+        std::thread::sleep(Duration::from_millis(50));
+    }
+    !process_is_alive(pid)
+}
+
+fn wait_for_orphaned_opencode_process(pid: u32, timeout: Duration) -> bool {
+    let deadline = Instant::now() + timeout;
+    while Instant::now() < deadline {
+        if is_orphaned_opencode_server_process(pid) {
+            return true;
+        }
+        std::thread::sleep(Duration::from_millis(50));
+    }
+    is_orphaned_opencode_server_process(pid)
+}
+
+fn build_service_with_store(
+    tasks: Vec<TaskCard>,
+    branches: Vec<GitBranch>,
+    current_branch: GitCurrentBranch,
+    config_store: AppConfigStore,
+) -> (AppService, Arc<Mutex<TaskStoreState>>, Arc<Mutex<GitState>>) {
+    let task_state = Arc::new(Mutex::new(TaskStoreState {
+        ensure_calls: Vec::new(),
+        ensure_error: None,
+        tasks,
+        list_error: None,
+        delete_calls: Vec::new(),
+        created_inputs: Vec::new(),
+        updated_patches: Vec::new(),
+        spec_get_calls: Vec::new(),
+        spec_set_calls: Vec::new(),
+        plan_get_calls: Vec::new(),
+        plan_set_calls: Vec::new(),
+        metadata_get_calls: Vec::new(),
+        qa_append_calls: Vec::new(),
+        latest_qa_report: None,
+        agent_sessions: Vec::new(),
+        upserted_sessions: Vec::new(),
+    }));
+    let git_state = Arc::new(Mutex::new(GitState {
+        calls: Vec::new(),
+        branches,
+        current_branch,
+    }));
+    let task_store: Arc<dyn TaskStore> = Arc::new(FakeTaskStore {
+        state: task_state.clone(),
+    });
+    let git_port: Arc<dyn GitPort> = Arc::new(FakeGitPort {
+        state: git_state.clone(),
+    });
+    let service = AppService::with_git_port(task_store, config_store, git_port);
+    (service, task_state, git_state)
+}
+
+fn make_emitter(events: Arc<Mutex<Vec<RunEvent>>>) -> Arc<dyn Fn(RunEvent) + Send + Sync> {
+    Arc::new(move |event| {
+        events.lock().expect("events lock poisoned").push(event);
+    })
+}
+
+fn spawn_sleep_process(seconds: u64) -> std::process::Child {
+    Command::new("/bin/sh")
+        .arg("-lc")
+        .arg(format!("sleep {seconds}"))
+        .stdout(Stdio::piped())
+        .stderr(Stdio::piped())
+        .spawn()
+        .expect("spawn sleep process")
+}
+
+fn empty_patch() -> UpdateTaskPatch {
+    UpdateTaskPatch {
+        title: None,
+        description: None,
+        acceptance_criteria: None,
+        notes: None,
+        status: None,
+        priority: None,
+        issue_type: None,
+        ai_review_enabled: None,
+        labels: None,
+        assignee: None,
+        parent_id: None,
+    }
+}
+
+mod integration;
+mod unit;
